# doc/api-spec.yaml
openapi: 3.0.0
info:
  title: Operation Center API
  description: Operation Center 系统的后端 API 接口文档
  version: 1.0.0
  contact:
    name: Operation Center Team
servers:
  - url: http://localhost:8080
    description: 本地开发服务器
paths:
...
  /api/find_setting:
    get:
      summary: 获取配置信息
      description: 通过GET方式获取指定目录中的配置文件内容
      operationId: getSettingsGet
      tags:
        - 配置管理
      parameters:
        - name: dir_path
          in: query
          required: true
          schema:
            type: string
          description: 解压后的目录路径
      responses:
        '200':
          description: 成功获取配置信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/save_setting:
    post:
      summary: 保存配置
      description: 将配置信息保存到指定目录
      operationId: saveSettings
      tags:
        - 配置管理
      parameters:
        - name: dir_path
          in: query
          required: true
          schema:
            type: string
          description: 解压后的目录路径
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveSettingsRequest'
      responses:
        '200':
          description: 配置保存成功
          content:
            text/plain:
              schema:
                type: string
                example: "配置保存成功"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'

  # 信息查询接口组
...
components:
  schemas:
...
    SaveSettingsRequest:
      type: object
      required:
        - settings
      properties:
        settings:
          $ref: '#/components/schemas/Settings'
          description: 配置项对象
    Settings:
      type: object
      description: 系统配置项
      properties:
        host:
          type: string
          description: 主机名
          example: "localhost"
        port:
          type: string
          description: 端口号
          example: "8080"
        username:
          type: string
          description: 用户名
          example: "admin"
        password:
          type: string
          description: 密码
          example: "******"
        db_name:
          type: string
          description: 数据库名称
          example: "operation_center"
        log_level:
          type: string
          description: 日志级别
          example: "info"
        auto_start:
          type: string
          description: 是否自动启动
          example: "true"
        backup_enabled:
          type: string
          description: 是否启用备份
          example: "true"
        max_log_size:
          type: string
          description: 最大日志大小
          example: "100MB"
        update_channel:
          type: string
          description: 更新渠道
          example: "stable"
      additionalProperties:
        type: string
        description: 其他配置项
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: 错误信息
        code:
          type: integer
          description: 错误代码
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "请求参数不正确"
            code: 400
    NotFound:
      description: 资源未找到
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "请求的资源不存在"
            code: 404
    ServerError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "服务器内部错误"
            code: 500
...